Move leveldb around
